PC3/requirements.txt
opencv-python==4.10.0.84
numpy==2.1.2
sounddevice==0.4.7
mysql-connector-python==8.2.0
python-dotenv==1.0.0
